requirements.txt
textblob==0.15.3
discord.py==1.2.2
pydub==0.23.1
discord==1.0.1
python-dotenv==0.10.3
PyNaCl==1.3.0
